[UPDATE] move train script to train_src and update gitignore
